pyproject.toml
[tool.poetry]
name = "project-1"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"
packages = [{include = "project_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
